hopefully fixed 2 bugs and made some additional small changes
BUG 1: occurred when clicking New and then adding a small # of points, occasionally the PathCalculators would calculate all paths before the PathJudge could run. This meant that the view would not be updated. To fix this, I moved the first part of the while condition to the end.

BUG 2: Was a NullPointerException thrown by viewport. For reasons I don't understand, sometimes a path in sortedPaths would be null. To hopefully fix this, each time getTopKPaths is called, all null objects are removed from sortedPaths

Other: moved notifyView to where it was intended to be in PathJudge. add a check in Repository for incorporating new points added after stopping.